rename "levelNumber" to "level"
